app/recommendation_system/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.4.2
pandas==2.1.2
numpy==1.24.3
scikit-learn==1.3.2
joblib==1.3.2
requests==2.31.0
beautifulsoup4==4.12.2
pymongo==4.5.0
python-dotenv==1.0.0
redis==5.0.1
pytest==7.4.3
prometheus-client==0.17.1
cachetools==5.3.2
psutil==5.9.6
memory-profiler==0.61.0 